feat: add new home page
