+++ src/lib/pages/home/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./DevShortcut";
+export * from "./TopDecorations";
