--- src/pages/[network]/network-overview.tsx
@@ -1,3 +1,0 @@
-import NetworkOverview from "lib/pages/network-overview";
-
-export default NetworkOverview;
