--- src/pages/dev-home.tsx
@@ -1,3 +1,0 @@
-import DevHome from "lib/pages/dev-home";
-
-export default DevHome;
